inv-475 Abstracting the inventory settings page and fixing the save button value in the setttings2x page
