+++ api/src/main/resources/messages.properties
@@ -113,9 +113,11 @@ openhmis.inventory.roleCreation.page.instruction=Please select a role or add a n
 openhmis.inventory.roleCreation.page.label.add=Add Inventory Privileges
 openhmis.inventory.roleCreation.page.label.remove=Remove Inventory Privileges
 openhmis.inventory.roleCreation.page.label.new=Create New Role
+openhmis.inventory.roleCreation.page.role.save=Save role
 
 # Settings
 openhmis.inventory.settings.saved=Settings Saved
+openhmis.inventory.settings.page.settings.save=Save Settings
 
 # Error
 openhmis.inventory.roleCreation.page.feedback.error.blankRole=Please Enter a Role Name
