+++ omod_1.x/src/main/webapp/roleCreation2x.jsp
@@ -100,6 +100,6 @@
 		</tr>
 	</table>
 
-	<p><input class="submitButton confirm right" value="Save" type="submit" value="<openmrs:message code="Role.save"/>"></p>
+	<p><input class="submitButton confirm right" type="submit" value="<openmrs:message code="openhmis.inventory.roleCreation.page.role.save"/>"></p>
 </form>
 </div>
